+++ R/get_statistic.R
@@ -416,18 +416,32 @@ get_statistic.Gam <- function(x, ...) {
 
 #' @export
 get_statistic.gam <- function(x, ...) {
-  cs <- summary(x)$p.table
-  cs.smooth <- summary(x)$s.table
+  p.table <- summary(x)$p.table
+  s.table <- summary(x)$s.table
 
-  out <- data.frame(
-    Parameter = c(rownames(cs), rownames(cs.smooth)),
-    Statistic = c(as.vector(cs[, 3]), as.vector(cs.smooth[, 3])),
-    Component = c(rep("conditional", nrow(cs)), rep("smooth_terms", nrow(cs.smooth))),
-    stringsAsFactors = FALSE,
-    row.names = NULL
-  )
+  d1 <- d2 <- NULL
 
-  out <- text_remove_backticks(out)
+  if (!is.null(p.table)) {
+    d1 <- data.frame(
+      Parameter = rownames(p.table),
+      Statistic = as.vector(p.table[, 3]),
+      Component = "conditional",
+      stringsAsFactors = FALSE,
+      row.names = NULL
+    )
+  }
+
+  if (!is.null(s.table)) {
+    d2 <- data.frame(
+      Parameter = rownames(s.table),
+      Statistic = as.vector(s.table[, 3]),
+      Component = "smooth_terms",
+      stringsAsFactors = FALSE,
+      row.names = NULL
+    )
+  }
+
+  out <- text_remove_backticks(rbind(d1, d2))
   attr(out, "statistic") <- find_statistic(x)
   out
 }
@@ -480,7 +494,7 @@ get_statistic.list <- function(x, ...) {
 #' @export
 get_statistic.gamlss <- function(x, ...) {
   parms <- get_parameters(x)
-  utils::capture.output(cs <- summary(x))
+  utils::capture.output(cs <- summary(x)) # nolint
 
   out <- data.frame(
     Parameter = parms$Parameter,
@@ -649,7 +663,7 @@ get_statistic.coxme <- function(x, ...) {
 
 #' @export
 get_statistic.riskRegression <- function(x, ...) {
-  junk <- utils::capture.output(cs <- stats::coef(x))
+  junk <- utils::capture.output(cs <- stats::coef(x)) # nolint
 
   out <- data.frame(
     Parameter = as.vector(cs[, 1]),
@@ -776,7 +790,7 @@ get_statistic.mvord <- function(x,
                                 component = c("all", "conditional", "thresholds", "correlation"),
                                 ...) {
   component <- match.arg(component)
-  junk <- utils::capture.output(s <- summary(x))
+  junk <- utils::capture.output(s <- summary(x)) # nolint
   # intercepts thresholds
   thresholds <- as.data.frame(s$thresholds)
   thresholds$Parameter <- rownames(thresholds)
@@ -1302,7 +1316,7 @@ get_statistic.ivprobit <- function(x, ...) {
 
 #' @export
 get_statistic.HLfit <- function(x, ...) {
-  utils::capture.output(s <- summary(x))
+  utils::capture.output(s <- summary(x)) # nolint
 
   out <- data.frame(
     Parameter = rownames(s$beta_table),
@@ -1662,7 +1676,7 @@ get_statistic.cpglm <- function(x, ...) {
   # installed?
   check_if_installed("cplm")
 
-  junk <- utils::capture.output(stats <- cplm::summary(x)$coefficients)
+  junk <- utils::capture.output(stats <- cplm::summary(x)$coefficients) # nolint
   params <- get_parameters(x)
 
   out <- data.frame(
@@ -1686,7 +1700,7 @@ get_statistic.zcpglm <- function(x,
   check_if_installed("cplm")
 
   component <- match.arg(component)
-  junk <- utils::capture.output(stats <- cplm::summary(x)$coefficients)
+  junk <- utils::capture.output(stats <- cplm::summary(x)$coefficients) # nolint
   params <- get_parameters(x)
 
   tweedie <- data.frame(
@@ -2086,7 +2100,7 @@ get_statistic.complmrob <- function(x, ...) {
 #' @export
 get_statistic.logistf <- function(x, ...) {
   parms <- get_parameters(x)
-  utils::capture.output(s <- summary(x))
+  utils::capture.output(s <- summary(x)) # nolint
 
   out <- data.frame(
     Parameter = parms$Parameter,
@@ -2203,7 +2217,7 @@ get_statistic.DirichletRegModel <- function(x,
                                             ...) {
   component <- match.arg(component)
   parms <- get_parameters(x)
-  junk <- utils::capture.output(cs <- summary(x)$coef.mat)
+  junk <- utils::capture.output(cs <- summary(x)$coef.mat) # nolint
 
   out <- data.frame(
     Parameter = parms$Parameter,