+++ NEWS.md
@@ -38,6 +38,12 @@
 
 * Fixes issue in `compact_list()`.
 
+* `has_single_value()` now returns `FALSE` when the object only has `NA` and 
+  `na.rm = TRUE`.
+
+* Fixed issue in `get_parameters()` for gam-models without smooth terms, or with
+  only smooth terms and removed intercept.
+
 # insight 0.18.8
 
 ## Bug fixes